+++ frontend/app/src/components/support-chat.tsx
