+++ frontend/app/src/pages/main/v1/workflow-runs-v1/$run/v2components/workflow-run-visualizer-v2.tsx
@@ -1,6 +1,6 @@
 import { useWorkflowDetails } from '../../hooks/use-workflow-details';
 import stepRunNode, { NodeData } from './step-run-node';
-import { useTheme } from '@/components/theme-provider';
+import { useTheme } from '@/components/hooks/use-theme';
 import { V1TaskStatus } from '@/lib/api';
 import dagre from 'dagre';
 import { useMemo } from 'react';
