+++ frontend/app/src/pages/onboarding/verify-email/index.tsx
@@ -1,4 +1,4 @@
-import MainNav from '@/components/molecules/nav-bar/nav-bar';
+import TopNav from '@/components/v1/nav/top-nav';
 import { Loading } from '@/components/v1/ui/loading';
 import api from '@/lib/api';
 import queryClient from '@/query-client';
@@ -47,7 +47,7 @@ export default function VerifyEmail() {
 
   return (
     <div className="flex h-full w-full flex-1 flex-row">
-      <MainNav user={res.user} tenantMemberships={[]} />
+      <TopNav user={res.user} tenantMemberships={[]} />
       <div className="container relative hidden flex-col items-center justify-center md:grid lg:max-w-none lg:grid-cols-2 lg:px-0">
         <div className="mx-auto w-screen lg:p-8">
           <div className="mx-auto flex w-40 flex-col justify-center space-y-6 sm:w-[350px]">
